move color disposition
