+++ packages/web/components/Player/Section/PlayerColorDisposition.tsx
@@ -1,8 +1,11 @@
-import { Wrap } from '@metafam/ds';
+import { Link } from '@metafam/ds';
 import { PlayerFragmentFragment } from 'graphql/autogen/types';
-import React from 'react';
+import { getPersonalityInfo } from 'graphql/getPersonalityInfo';
+import { PersonalityOption } from 'graphql/types';
+import React, { useEffect } from 'react';
 
 import { ProfileSection } from '../../ProfileSection';
+import { ColorBar } from '../ColorBar';
 
 type Props = {
   player: PlayerFragmentFragment;
@@ -13,12 +16,38 @@ export const PlayerColorDisposition: React.FC<Props> = ({
   player,
   displayEditButton,
   onRemoveClick,
-}) => (
-  <ProfileSection
-    title="Color Disposition"
-    onRemoveClick={onRemoveClick}
-    displayEditButton={displayEditButton}
-  >
-    <Wrap />
-  </ProfileSection>
-);
+}) => {
+  const [types, setTypes] = React.useState<{
+    [any: string]: PersonalityOption;
+  }>();
+  const mask = player?.color_aspect?.mask;
+  const type = mask && types?.[mask];
+
+  const loadTypes = async () => {
+    const { types: list } = await getPersonalityInfo();
+    setTypes(list);
+  };
+  useEffect(() => {
+    loadTypes();
+  }, []);
+
+  return (
+    <ProfileSection
+      title="Color Disposition"
+      onRemoveClick={onRemoveClick}
+      displayEditButton={displayEditButton}
+    >
+      {type && types && (
+        <Link
+          isExternal
+          href={`//dysbulic.github.io/5-color-radar/#/combos/${type.mask.toString(
+            2,
+          )}`}
+          maxH="6rem"
+        >
+          <ColorBar mask={type.mask} />
+        </Link>
+      )}
+    </ProfileSection>
+  );
+};
