+++ packages/web/components/Player/Section/PlayerHero.tsx
@@ -4,20 +4,16 @@ import {
   Flex,
   HStack,
   IconButton,
-  Link,
   Text,
   VStack,
 } from '@metafam/ds';
 import { PlayerAvatar } from 'components/Player/PlayerAvatar';
 import { PlayerFragmentFragment } from 'graphql/autogen/types';
-import { getPersonalityInfo } from 'graphql/queries/enums/getPersonalityInfo';
-import { PersonalityOption } from 'graphql/types';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { getPlayerDescription, getPlayerName } from 'utils/playerHelpers';
 
 import { FlexContainer } from '../../Container';
 import { ProfileSection } from '../../ProfileSection';
-import { ColorBar } from '../ColorBar';
 import { PlayerContacts } from '../PlayerContacts';
 import { PlayerBrightId } from './PlayerBrightId';
 import { PlayerCollab } from './PlayerCollab';
@@ -28,19 +24,6 @@ type Props = { player: PlayerFragmentFragment; isOwnProfile: boolean };
 export const PlayerHero: React.FC<Props> = ({ player, isOwnProfile }) => {
   const description = getPlayerDescription(player);
   const [show, setShow] = React.useState(description.length <= MAX_BIO_LENGTH);
-  const [types, setTypes] = React.useState<{
-    [any: string]: PersonalityOption;
-  }>();
-  const mask = player?.color_aspect?.mask;
-  const type = mask && types?.[mask];
-
-  const loadTypes = async () => {
-    const { types: list } = await getPersonalityInfo();
-    setTypes(list);
-  };
-  useEffect(() => {
-    loadTypes();
-  }, []);
 
   return (
     <ProfileSection>
@@ -98,31 +81,6 @@ export const PlayerHero: React.FC<Props> = ({ player, isOwnProfile }) => {
         <Box w="100%">
           <PlayerCollab player={player} />
         </Box>
-        {type && types && (
-          <Flex direction="column" id="color" mb={0} w="100%">
-            <Text
-              fontSize="xs"
-              color="blueLight"
-              casing="uppercase"
-              mb={3}
-              textAlign="left"
-            >
-              Color Disposition
-            </Text>
-            <Link
-              isExternal
-              href={`//dysbulic.github.io/5-color-radar/#/combos/${type.mask.toString(
-                2,
-              )}`}
-              maxH="6rem"
-            >
-              <ColorBar mask={type.mask} />
-            </Link>
-            <Text color="blueLight" mt={4} style={{ textIndent: 16 }}>
-              {type.description}
-            </Text>
-          </Flex>
-        )}
         {player.type?.title && (
           <FlexContainer align="stretch" fontSize={{ base: 'sm', sm: 'md' }}>
             <Text color="white" fontWeight="bold" casing="uppercase">
